Finally fixed the biiiiiig leak
This took entirely too long. I was creating two new buffer objects and an array buffer object every frame and never destroying them. Fixed that. Now the only memory leak I know of is the vector3f leak that'll get fixed with a QuadRendererComponent class rewrite. That class is a pile of crap anyway. Anyway, those vectors get GC'd so it's not that big of a deal but there will be performance issues on it down the line in big scenes so I should just clean it up now.