Got textures working, massively updated flappy bird demo
Still not sure what the tiny stutter is caused by.

Enabled depth testing.

Now have a textured and non-textured shader, and textured and non-textured mesh renderers know which one they own.

Now able to keep track of textures and shaders of items in render queue. Not particularly elegant, but works well enough that I'm not going to touch it again.