Took a week off, made stuff work
Was very annoying before, took a week off and when I came back I realized my game loop flow was needlessly complex and simplified it. Also added a priority execution function for any GameObjectComponent. This is useful, for instance, when a sound is created and played in the same function (i.e. in the same onCreate). It's also useful for re-running a single component's onCreate() function.