Cross-platform file loading, fixed a little issue where scenes weren't getting cleared on unload.
Don't like the way I'm handling scenes, and if I want to serialize, I'll need to change the way it works a bit.

Also I still don't have a code-reuse function for file loading, but I'm only loading from file in two places at the moment so I'm not fussed about it.